oidc_provider/requirements.txt
pyjwt==2.9.0
cryptography==43.0.1
appdirs==1.4.4
backcall==0.2.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
dnspython==2.0.0
ecdsa==0.16.0
email-validator==1.1.1
fastecdsa==2.1.5
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.9
Flask-Script==2.0.6
flask-shell-ipython==0.4.1
gevent==20.9.0
greenlet==0.4.17
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
ipaddr==2.2.0
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.4
parso==0.7.1
pep517==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
progress==1.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
py-healthcheck==1.10.1
Pygments==2.7.1
pyparsing==2.4.7
python-engineio==3.13.2
python-usernames==0.3.1
pytoml==0.1.21
qrcode==6.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.3.1
psycopg2-binary==2.8.6
pygit2==1.4.0
python-dotenv==1.0.0